Add more generics
