+++ examples/aggregator/client/aggregator.cc
@@ -21,7 +21,6 @@
 #include "examples/aggregator/proto/aggregator.pb.h"
 #include "include/grpcpp/grpcpp.h"
 #include "oak/client/application_client.h"
-#include "oak/common/nonce_generator.h"
 
 ABSL_FLAG(std::string, address, "127.0.0.1:8080", "Address of the Oak application to connect to");
